Added SageMaker batch transform support.
